+++ lib/main_db/new/models/session.dart
@@ -1,0 +1,8 @@
+import 'package:hoplixi/main_db/core/main_store.dart';
+import 'package:hoplixi/main_db/core/models/dto/index.dart';
+
+typedef Session = ({
+  MainStore store,
+  StoreInfoDto info,
+  String storeDirectoryPath,
+});