+++ lib/main_db/new/usecases/close_main_store.dart
@@ -1,7 +1,6 @@
 import 'package:hoplixi/core/errors/errors.dart';
 import 'package:hoplixi/core/logger/index.dart' hide Session;
-import 'package:hoplixi/main_db/new/usecases/create_main_store.dart'
-    show Session;
+import 'package:hoplixi/main_db/new/models/session.dart';
 import 'package:hoplixi/main_db/new/usecases/utils/error_handling.dart';
 import 'package:result_dart/result_dart.dart';