+++ lib/main_db/new/usecases/update_main_store.dart
@@ -5,6 +5,7 @@ import 'package:drift/drift.dart';
 import 'package:hoplixi/core/errors/errors.dart';
 import 'package:hoplixi/core/logger/index.dart' hide Session;
 import 'package:hoplixi/main_db/core/models/dto/index.dart';
+import 'package:hoplixi/main_db/new/models/session.dart';
 import 'package:hoplixi/main_db/new/usecases/create_main_store.dart'
     show Session;
 import 'package:hoplixi/main_db/new/usecases/utils/error_handling.dart';